Made automation changes only affect the current ControlChange. Made use of move/edit/delete buttons on automation graph. Cleaned up some warnings.
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@1512 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/editmidi/MidiPart.cpp
@@ -359,7 +359,7 @@ void					MidiPart::OnClick(wxMouseEvent &e)
 	  {
 	    if (e.GetX() - b.x > 0)
 	    {
-	      long duration = (e.GetX() - b.x) / (ROW_WIDTH * 4 * ZoomX);
+	      long duration = (long)((e.GetX() - b.x) / (ROW_WIDTH * 4 * ZoomX));
 	      selected2->SetDuration(duration);
 	      Refresh(true);
 	      em->ma->SetNotes(Notes);
